Add support for saving scripts.
Add back and front end support for saving script to Cloud Datastore.
Add button to front end to save script, a request is posted to the back
end and a unique ID is returned if the save is succesful, the user is
then redirected to the URL of their new script.