Start work on mini editor. Add skeleton html + ts
This commit starts work on the mini editor. Specifically adding a
`mini-editor` custom element with the basic structure.

The mini-editor will be incorporated into a series of interactive
tutorials. It's to be small enough that it can be inlined into a
tutorial site several times, with runnable examples of code or "fill in
the blanks" exercises.