Move `luch_síos` to `luch` in `stáitse` object
We rename `luch_síos` to `luch` to be more inline with the naming of
`méarchlár`. Future plans are to add `luch_suas` and `méarchlár_suas`
for the mouse up and key up events respectively.

We move the `luch_síos` action into the `stáitse` object for consistency
as other stage-specific actions are in this object.