Add fán (wait) action for infinite sleep
Lots of Setanta programs rely on event driven concurrency so
need to add some sort of infinite loop at the end of the main
thread to stop the program from assigning event handlers and
immmediately exiting. These loops are inefficient and eat CPU cycles.

Fán is an action that will permanently pause the thread it is executed
in. It does this efficiently.

Fán will be absorbed into the Setanta interpreter builtins in the future
but for now is only available in browser.